+++ cmd/APK.go
@@ -3,6 +3,8 @@ package cmd
 import (
 	"errors"
 	"fmt"
+	"os/exec"
+	"strings"
 
 	"github.com/bitrise-io/go-utils/log"
 )
@@ -11,7 +13,6 @@ import (
 func APK(file, path string, args []string) error {
 	log.Infof("Retrieving APK Info:")
 	// file and path is redundant if you work with args as well
-	printInfo(file, path)
 	if len(args) > 0 {
 		appInfo, err := getAPK(args[0])
 		if err != nil {
@@ -21,4 +22,23 @@ func APK(file, path string, args []string) error {
 		return nil
 	}
 	return errors.New("Index out of bounds")
+}
+
+func getAPK(path string) (map[string]string, error) {
+	command, err := exec.Command("aapt", "dump", "badging", path).Output() // works only if aapt is in PATH, see: https://github.com/bitrise-io/steps-deploy-to-bitrise-io/blob/master/uploaders/apkuploader.go#L69-L82
+	appInfo := make(map[string]string)
+	if err != nil {
+		return appInfo, err
+	}
+	info := strings.Split(string(command), "\n")
+	for i := 0; i < len(info); i++ { // for _, line := range info {}
+		if strings.Contains(info[i], "package") {
+			appInfo["packageName"] = strings.Split(info[i], "'")[1]
+			appInfo["versionCode"] = strings.Split(info[i], "'")[3]
+			appInfo["versionName"] = strings.Split(info[i], "'")[5]
+		} else if strings.Contains(info[i], "application-icon") {
+			appInfo["icon"] = strings.Split(info[i], "'")[1]
+		}
+	}
+	return appInfo, nil
 }
